Move Geometry to GPlacesPhp\ApiClient\Client namespace.
