+++ tests/Unit/Client/Geometry/LocationTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GPlacesPhp\ApiClient\Tests\Unit\Client\PlaceDetails\Geometry;
+namespace GPlacesPhp\ApiClient\Tests\Unit\Client\Geometry;
 
 use GPlacesPhp\ApiClient\Tests\TestCase\Mother\LocationMother;
 use PHPUnit\Framework\TestCase;
